Give most classes `__slots__` and move somes files around for cleanliness.
Using `__slots__` will make it easier to serialize objects.
Also, allow `ParameterizedBinaryOp` to be associated with a monoid.
