+++ grblas/ss/_utils.pxd
