+++ grblas/ss/_utils.pyx
